// comunicacionesDirectas/src/shared/repositories/autogenerado/wsMensajeriaClient.ts
export interface IComunicacionDirectaClient {
  /**
   * @param x_api_version (optional)
   * @param body (optional)
   * @return Created
   */
  enviarMensaje(
    x_api_version?: string | undefined,
    body?: EnviarMensajeCommand | undefined,
    signal?: AbortSignal,
  ): Promise<string>;
  /**
   * @param x_api_version (optional)
   * @param body (optional)
   * @return Created
   */
  archivarComunicacionesDirectas(
    x_api_version?: string | undefined,
    body?: ArchivarConversacionesCommand | undefined,
    signal?: AbortSignal,
  ): Promise<void>;
  /**
   * @param x_api_version (optional)
   * @param body (optional)
   * @return Created
   */
  actualizarConsentimientoComunicaciones(
    x_api_version?: string | undefined,
    body?: ActualizarConsentimientoComunicacionesCommand | undefined,
    signal?: AbortSignal,
  ): Promise<void>;
  /**
   * @param request (optional)
   * @param x_api_version (optional)
   * @return OK
   */
  comprobarConsentimientoComunicaciones(
    request?: ComprobarConsentimientoComunicacionesQuery | undefined,
    x_api_version?: string | undefined,
    signal?: AbortSignal,
  ): Promise<boolean>;
  /**
   * @param activas (optional)
   * @param x_api_version (optional)
   * @return OK
   */
  obtenerConversacionesPorUsuario(
    activas?: boolean | undefined,
    x_api_version?: string | undefined,
    signal?: AbortSignal,
  ): Promise<ConversacionesPorUsuarioResponse>;
  /**
   * @param idConversacionEncriptado (optional)
   * @param x_api_version (optional)
   * @return OK
   */
  obtenerMensajesPorConversacion(
    idConversacionEncriptado?: string | undefined,
    x_api_version?: string | undefined,
    signal?: AbortSignal,
  ): Promise<MensajesPorConversacionResponse>;
  /**
   * @param request (optional)
   * @param x_api_version (optional)
   * @return OK
   */
  obtenerConversacionesConMensajesNuevos(
    request?: ObtenerConversacionesConMensajesNuevosQuery | undefined,
    x_api_version?: string | undefined,
    signal?: AbortSignal,
  ): Promise<string[]>;
  /**
   * @param request (optional)
   * @param x_api_version (optional)
   * @return OK
   */
  obtenerArchivadasConMensajesNuevos(
    request?: ObtenerArchivadasConMensajesNuevosQuery | undefined,
    x_api_version?: string | undefined,
    signal?: AbortSignal,
  ): Promise<number>;
  /**
   * @param x_api_version (optional)
   * @param body (optional)
   * @return Created
   */
  marcarMensajesLeidos(
    x_api_version?: string | undefined,
    body?: MarcarMensajesLeidosCommand | undefined,
    signal?: AbortSignal,
  ): Promise<void>;
  /**
   * @param x_api_version (optional)
   * @param body (optional)
   * @return Created
   */
  asignarArchivoAMensaje(
    x_api_version?: string | undefined,
    body?: AsignarArchivoAMensajeCommand | undefined,
    signal?: AbortSignal,
  ): Promise<void>;
}
// ...
export interface ActualizarConsentimientoComunicacionesCommand {
  consentimiento?: boolean;
}

export interface ArchivarComunicacionesDirectasPorUsuarioCommand {
  idsUsuariosEncriptados?: string[] | undefined;
}

export interface ArchivarConversacionesCommand {
  idsComunicacionesEncriptados?: string[] | undefined;
}

export interface AsignarArchivoAMensajeCommand {
  idComunicacionEncriptado?: string | undefined;
  idArchivo?: string | undefined;
  idMensaje?: string | undefined;
}

export interface ComprobarComunicacionesActivadasQuery {}

export interface ComprobarConsentimientoComunicacionesQuery {}

export interface Conversacion {
  idConversacionEncriptado?: string | undefined;
  titulo?: string | undefined;
  nombreParticipante?: string | undefined;
  ultimoMensaje?: string | undefined;
  fechaUltimoMensaje?: Date;
  estadoUltimoMensaje?: number;
  tipoUltimoMensaje?: number;
  esPropioMensaje?: boolean;
  mensajesSinLeer?: number;
}

export interface ConversacionesPorUsuarioResponse {
  conversaciones?: Conversacion[] | undefined;
}

export interface CrearComunicacionDirectaConElMagistradoCommand {
  idOrganoJudicial?: number;
  titulo?: string | undefined;
  contenidoMensaje?: string | undefined;
  idArchivoEncriptado?: string | undefined;
}

export interface EnviarMensajeCommand {
  idComunicacionEncriptado?: string | undefined;
  contenidoMensaje?: string | undefined;
}

export interface MarcarMensajesLeidosCommand {
  idsMensajesLeidosEncriptados?: string[] | undefined;
}

export interface Mensaje {
  tipoMensaje?: number;
  texto?: string | undefined;
  localizadorArchivo?: string | undefined;
  fechaEnvio?: Date;
  esPropioMensaje?: boolean;
  idMensajeEncriptado?: string | undefined;
  estadoMensaje?: number;
}

export interface MensajesPorConversacionResponse {
  idConversacionEncriptado?: string | undefined;
  idEstadoConversacion?: number;
  usuarioRemitente?: string | undefined;
  mensajes?: Mensaje[] | undefined;
}

export interface ObtenerArchivadasConMensajesNuevosQuery {}

export interface ObtenerConversacionesConMensajesNuevosQuery {}